// src/app/model/assessment_4.ts
export class Assessment4 {
    t1: number = 0;
    t2: number = 0;
    t3: number = 0;
    t4: number = 0;
    t5: number = 0;
    t6: number = 0;
    t7: number = 0;
    t8: number = 0;
    t9: number = 0;
    t10: number = 0;
    t11: number = 0;
    t12: number = 0;
    t13: number = 0;
    t14: number = 0;
    t15: number = 0;
    t16: number = 0;
    t17: number = 0;
    s1: string;
    s2: string;
    s3: string;
    s4: string;
    s5: string;
    s6: string;
    s7: string;
    s8: string;
    s9: string;
    s10: string;
    s11: string;
    s12: string;
    s13: string;
    s14: string;
    s15: string;
    s16: string;
    s17: string;
}